deleted filestore.go and moved buildPath()
